Start moving database provider code into SecureStorage.
